+++ src/managers/UnlockPanelManager/UnlockPanelManager.ts
@@ -9,7 +9,7 @@ import {
 } from 'providers/types/providerFactory.types';
 import {
   AllowedProviderType,
-  CloseCallbackType,
+  OnCloseUnlockPanelType,
   LoginHandlerType,
   UnlockPanelEventsEnum,
   UnlockPanelManagerInitParamsType
@@ -25,7 +25,7 @@ export class UnlockPanelManager extends SidePanelBaseManager<
 
   private static instance: UnlockPanelManager;
   private static loginHandler: LoginHandlerType | null = null;
-  private static closeCallback: CloseCallbackType | null = null;
+  private static onClose: OnCloseUnlockPanelType | null = null;
   private static allowedProviders?: AllowedProviderType[] | null = null;
 
   constructor() {
@@ -48,8 +48,8 @@ export class UnlockPanelManager extends SidePanelBaseManager<
     this.loginHandler = params.loginHandler;
     this.allowedProviders = params.allowedProviders;
 
-    if (params.closeCallback) {
-      this.closeCallback = params.closeCallback;
+    if (params.onClose) {
+      this.onClose = params.onClose;
     }
 
     return this.getInstance();
@@ -74,15 +74,17 @@ export class UnlockPanelManager extends SidePanelBaseManager<
     this.subscribeToEventBus(UnlockPanelEventsEnum.CLOSE, this.handleCloseUI);
   };
 
-  private handleCloseUI = async (options?: { isLoginFinished?: boolean }) => {
-    if (!options?.isLoginFinished && UnlockPanelManager.closeCallback) {
-      UnlockPanelManager.closeCallback();
+  private readonly handleCloseUI = async (options?: {
+    isLoginFinished?: boolean;
+  }) => {
+    if (!options?.isLoginFinished && UnlockPanelManager.onClose) {
+      UnlockPanelManager.onClose();
     }
 
     this.closeUI();
   };
 
-  private handleLogin = async ({ type, anchor }: IProviderFactory) => {
+  private readonly handleLogin = async ({ type, anchor }: IProviderFactory) => {
     if (!UnlockPanelManager.loginHandler) {
       throw new Error(
         'Login callback not initialized. Please call `init()` first.'
@@ -107,11 +109,11 @@ export class UnlockPanelManager extends SidePanelBaseManager<
     }
   };
 
-  private handleCancelLogin = async () => {
+  private readonly handleCancelLogin = async () => {
     await ProviderFactory.destroy();
   };
 
-  private isSimpleLoginCallback = (
+  private readonly isSimpleLoginCallback = (
     login: LoginHandlerType
   ): login is () => void => {
     const takesZeroArguments = login.length === 0;