+++ src/methods/network/getExplorerAddress.ts
@@ -1,4 +1,4 @@
-import { networkSelector } from 'store/selectors';
+import { networkSelector } from 'store/selectors/networkSelectors';
 import { getState } from 'store/store';
 
 export function getExplorerAddress(state = getState()) {